requirements-railway.txt
# Lightweight requirements for Railway deployment
# Core web framework
fastapi
uvicorn[standard]==0.24.0
pydantic
python-multipart==0.0.6

# Basic data processing (much lighter than full ML stack)
scikit-learn>=1.4.0
numpy>=1.24.0

# HTTP and utilities
requests>=2.28.0
python-dotenv>=1.0.0

# OpenAI API (lightweight)
openai>=1.12.0

# Note: ChromaDB and sentence-transformers removed for Railway deployment
# The app will use pre-computed embeddings and TF-IDF fallback only